Bug: Data is not cleaning properly
OS: Window

no. of camera: 10

Version: 1.0.0  Testing: data cleaning

expected result: after the file is recorded successfully to the wasabi
file must be deleted

Actual result: A few files are still there in the temp folder

root cause while writing file multiple time it is causing issue to close
the file delayed causing file to remove.

solution: added timer of 5 minutes delete the files that were not
able to be removed before to delete it